refactoring (remove redux-zero)
