+++ TODO.md
@@ -1,0 +1,14 @@
+# TODO
+
+#### Tasks
+
+- [x] [fe] parent-child task presentation
+- [ ] [fe] drag-and-drop ordering
+- [ ] [fe] drag-and-drop parent-child
+- [ ] [fe] allow collapsing of child tasks
+    - [ ] store collapsing state in DB (or locally?)
+- [ ] [fe-test] drag-and-drop logic
+- [ ] [fe-test] parent-child logic
+- [ ] [fe-test] parent-child collapsing logic
+- [ ] [be] handle ordering
+- [ ] [be] handle parent-child relationships
