--- web-client/src/redux/tasks.ts
@@ -1,3 +1,0 @@
-// TODO: set up redux store for tasks
-import * as createStore from 'redux-zero';
-import { Provider, connect } from 'redux-zero/preact';
