+++ web-client/tests/fixtures/tasks.ts
@@ -1,0 +1,64 @@
+import { APITask } from '../../src/tasks/TaskList';
+
+export const API_TASKS: Array<APITask> = [
+    {
+        id: 'task_abc123',
+        position: 1,
+        description: 'task one',
+        completed: false,
+        tags: [],
+        resources: [],
+    },
+    {
+        id: 'task_def456',
+        position: 2,
+        description: 'task two',
+        completed: false,
+        tags: [],
+        resources: [],
+    },
+    {
+        id: 'task_ghi789',
+        parentId: 'task_def456',
+        position: 1,
+        description: 'task three',
+        completed: true,
+        tags: [],
+        resources: [],
+    },
+    {
+        id: 'task_ghi238',
+        parentId: 'task_ghi789',
+        position: 2,
+        description: 'sub-task one',
+        completed: false,
+        tags: [],
+        resources: [],
+    },
+    {
+        id: 'task_ghi369',
+        parentId: 'task_ghi789',
+        position: 3,
+        description: 'sub-task two',
+        completed: false,
+        tags: [],
+        resources: [],
+    },
+    {
+        id: 'task_ghi723',
+        parentId: 'task_def456',
+        position: 1,
+        description: 'task 3b',
+        completed: false,
+        tags: [],
+        resources: [],
+    },
+    {
+        id: 'task_xyz987',
+        position: 3,
+        description: 'task four',
+        completed: false,
+        tags: [],
+        resources: [],
+    },
+];
